MB-42577 Properly handle selectivities of LIKE predicates
LIKE predicates with leading wildcard cannot be costed properly.
It'll have a default selectivity. For index scan such LIKE
predicate can generate a span that includes all strings.
Currently there is no way to link such index span with the LIKE
predicate and thus it's not clear that the selectivity may have
been applied already.

Add new flags to mark the index span to indicate it's for LIKE
predicate with default selectivity. Then when marking index
filters also find and mark the LIKE predicate such that its
selectivity is not reapplied.

Also fixed an issue with early filtering where we apply the
transformed LIKE predicate but then remove the original LIKE
predicate from the FILTER operator later on. This is not always
correct. Instead we only remove the LIKE predicate if the
original LIKE predicate is being applied early.

Reviewed-on: http://review.couchbase.org/c/query/+/142435
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>

+++ expression/expression.go
@@ -206,17 +206,17 @@ type Expression interface {
 	/*
 	   Check expression flag
 	*/
-	HasExprFlag(flag uint32) bool
+	HasExprFlag(flag uint64) bool
 
 	/*
 	   Set expression flag
 	*/
-	SetExprFlag(flag uint32)
+	SetExprFlag(flag uint64)
 
 	/*
 	   Unset expression flag
 	*/
-	UnsetExprFlag(flag uint32)
+	UnsetExprFlag(flag uint64)
 
 	/*
 	   Enable in-list evaluation optimization (using hash table)
